% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armIndello.R
\name{armIndello}
\alias{armIndello}
\alias{armIndello,XString,GRanges,GRanges-method}
\alias{armIndello,XStringSet,GRanges,GRanges-method}
\alias{armIndello,DNAStringSet,GRanges,missing-method}
\alias{armIndello,BSgenome,GRanges,missing-method}
\title{Substitute InDels into one or more sequences}
\usage{
armIndello(x, indels, exons, ...)

\S4method{armIndello}{XString,GRanges,GRanges}(x, indels, exons, alt_col = "ALT", ...)

\S4method{armIndello}{XStringSet,GRanges,GRanges}(x, indels, exons, alt_col = "ALT", ...)

\S4method{armIndello}{DNAStringSet,GRanges,missing}(x, indels, exons, alt_col = "ALT", mc_cores = 1, ...)

\S4method{armIndello}{BSgenome,GRanges,missing}(x, indels, exons, alt_col = "ALT", mc_cores = 1, ...)
}
\arguments{
\item{x}{Sequence of class XString}

\item{indels}{GRanges object with InDel locations and the alternate allele}

\item{exons}{GRanges object containing exon structure for \code{x}}

\item{...}{Not used}

\item{alt_col}{Column containing the alternate allele}

\item{mc_cores}{Number of cores to use when calling \link[parallel:mclapply]{parallel::mclapply}
internally}
}
\description{
Modify one or more sequences to include Insertions or Deletions
}
\details{
Takes an \link[Biostrings:XString-class]{Biostrings::XString} or \link[Biostrings:XStringSet-class]{Biostrings::XStringSet} object and moddfies
the sequence to incorporate InDels.
The expected types of data determine the behaviour, with the following
expectations describing how the function will incorporate data\tabular{llll}{
   Input Data Type \tab Exons Required \tab Use Case \tab Returned \cr
   XString \tab Y \tab Modify a Reference Transcriptome \tab XString \cr
   XStringSet \tab Y \tab Modify a Reference Transcriptome \tab XStringSet \cr
   DNAStringSet \tab N \tab Modify a Reference Genome \tab DNAStringSet \cr
   BSgenome \tab N \tab Modify a Reference Genome \tab DNAStringSet \cr
}
}
