% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoMog.R
\name{genoMog}
\alias{genoMog}
\alias{genoMog,XStringSet,GRanges-method}
\alias{genoMog-methods}
\alias{genoMog,BSgenome,GRanges-method}
\title{Mogrify a genome using a set of variants}
\usage{
genoMog(x, var, ...)

\S4method{genoMog}{XStringSet,GRanges}(x, var, alt_col = "ALT", seq, tag = NULL, sep = "_", ...)

\S4method{genoMog}{BSgenome,GRanges}(x, var, alt_col = "ALT", seq, tag = NULL, sep = "_", ...)
}
\arguments{
\item{x}{A DNAStringSet or BSgenome}

\item{var}{GRanges object containing the variants}

\item{...}{Passed to armIndello}

\item{alt_col}{The name of the column with \code{var} containing alternate bases}

\item{seq}{Optional subset of sequences to apply the function to}

\item{tag}{Optional tag to add to all sequence names}

\item{sep}{Separator to place between seqnames names & tag}
}
\value{
XStringSet with variant modified sequences
}
\description{
Use a set of insertions and deletions to modify a reference genome
}
\examples{
library(GenomicRanges)
seq <- DNAStringSet(c(chr1 = "ACGT"))
var <- GRanges(c("chr1:1", "chr1:3"))
var$ALT <- c("C", "GG")
seq
genoMog(seq, var, tag = "mod")


}
