% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mogrifyGenome.R
\name{mogrifyGenome}
\alias{mogrifyGenome}
\alias{mogrifyGenome,XStringSet,GRanges-method}
\alias{mogrifyGenome-methods}
\alias{mogrifyGenome,BSgenome,GRanges-method}
\title{Mogrify a genome using a set of variants}
\usage{
mogrifyGenome(x, var, ...)

\S4method{mogrifyGenome}{XStringSet,GRanges}(
  x,
  var,
  alt_col = "ALT",
  tag = NULL,
  sep = "_",
  var_tags = FALSE,
  var_sep = "_",
  ...
)

\S4method{mogrifyGenome}{BSgenome,GRanges}(
  x,
  var,
  alt_col = "ALT",
  names,
  tag = NULL,
  sep = "_",
  var_tags = FALSE,
  var_sep = "_",
  ...
)
}
\arguments{
\item{x}{A DNAStringSet or BSgenome}

\item{var}{GRanges object containing the variants}

\item{...}{Passed to armIndello}

\item{alt_col}{The name of the column with \code{var} containing alternate bases}

\item{tag}{Optional tag to add to all sequence names which were modified}

\item{sep}{Separator to place between seqnames names & tag}

\item{var_tags}{logical(1) Add tags indicating which type of variant were
incorporated, with 's', 'i' and 'd' represeenting SNPs, Insertions and
Deletions respectively}

\item{var_sep}{Separator between any previous tags and variant tags}

\item{names}{Sequence names to be mogrified}
}
\value{
XStringSet with variant modified sequences
}
\description{
Use a set of SNPS, insertions and deletions to modify a reference genome
}
\examples{
library(GenomicRanges)
dna <- DNAStringSet(c(chr1 = "ACGT", chr2 = "AATTT"))
var <- GRanges(c("chr1:1", "chr1:3", "chr2:1-3"))
var$ALT <- c("C", "GG", "A")
dna
mogrifyGenome(dna, var)
mogrifyGenome(dna, var, tag = "mod")
mogrifyGenome(dna, var, var_tags = TRUE)


}
