% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapsByVar.R
\name{overlapsByVar}
\alias{overlapsByVar}
\title{Count overlaps by variant type}
\usage{
overlapsByVar(x, var, alt_col = "ALT", ...)
}
\arguments{
\item{x}{A GRangesList with features of interest}

\item{var}{A Granges object with variants of interest}

\item{alt_col}{The column within mcols(var) which contains the alternate
allele}

\item{...}{Passed to \link{rowSums}}
}
\value{
A data.frame
}
\description{
Count how many variants overlap each element of a GRangesList
}
\details{
Taking any GRangesList, count how many of each variant type overlap a region
within each element.
}
\examples{
library(rtracklayer)
library(VariantAnnotation)
gtf <- import.gff(
    system.file("extdata/gencode.v44.subset.gtf.gz", package = "transmogR")
)
grl <- splitAsList(gtf, gtf$type)
vcf <- system.file("extdata/1000GP_subset.vcf.gz", package = "transmogR")
var <- rowRanges(readVcf(vcf, param = ScanVcfParam(fixed = "ALT")))
overlapsByVar(grl, var)

}
