% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parFromSeqinfo.R
\name{parFromSeqinfo}
\alias{parFromSeqinfo}
\title{Get the PAR-Y Regions From a Seqinfo Object}
\usage{
parFromSeqinfo(x, ...)
}
\arguments{
\item{x}{A Seqinfo object}

\item{...}{Not used}
}
\description{
Define the Pseudo-Autosomal Regions from a Seqinfo Object
}
\details{
Using a seqinfo object based on either hg38, hg19 or their variations,
create a GRanges object with the Pseudo-Autosomal Regions from the Y
chromosome for that build.

An additional mcols column called PAR will indicate PAR1 and PAR2
}
\examples{
library(GenomeInfoDb)
sq <- Seqinfo(
  seqnames = "chrY", seqlengths = 59373566, genome = "hg19_only_chrY"
)
parFromSeqinfo(sq)


}
\references{
GenomicRanges object
}
