% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftByVar.R
\name{shiftByVar}
\alias{shiftByVar}
\title{Calculate new exon co-ordinates}
\usage{
shiftByVar(x, var, alt_col = "ALT", mc.cores = 1, ...)
}
\arguments{
\item{x}{A GenomicRanges object with co-ordinates needing to be recalculated}

\item{var}{A set of variants to be incorporated into a reference genome}

\item{alt_col}{The name of the column with the alternate sequence for each
variant}

\item{mc.cores}{Passed internally to \code{\link[parallel:mclapply]{parallel::mclapply()}}}

\item{...}{Not used}
}
\description{
Calcluate new exon co-ordinates after including InDels
}
\details{
Given a set of variants, this will return a set of genomic ranges with
updated co-ordinates able to be applied on a variant modified reference
genome
}
