% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subInDel.R
\name{subInDel}
\alias{subInDel}
\alias{subInDel,XString,GRanges,GRanges-method}
\alias{subInDel,XStringSet,GRanges,GRanges-method}
\alias{subInDel,DNAStringSet,GRanges,missing-method}
\alias{subInDel,BSgenome,GRanges,missing-method}
\title{Incorporate InDels into one or more sequences}
\usage{
subInDel(x, indels, exons, ...)

\S4method{subInDel}{XString,GRanges,GRanges}(x, indels, exons, alt_col = "ALT", ...)

\S4method{subInDel}{XStringSet,GRanges,GRanges}(x, indels, exons, alt_col = "ALT", ...)

\S4method{subInDel}{DNAStringSet,GRanges,missing}(x, indels, exons, alt_col = "ALT", ..., BPPARAM = SerialParam())

\S4method{subInDel}{BSgenome,GRanges,missing}(x, indels, exons, alt_col = "ALT", ..., BPPARAM = SerialParam())
}
\arguments{
\item{x}{Sequence of class XString}

\item{indels}{GRanges object with InDel locations and the alternate allele}

\item{exons}{GRanges object containing exon structure for \code{x}}

\item{...}{Not used}

\item{alt_col}{Column containing the alternate allele}

\item{BPPARAM}{\link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} instance}
}
\description{
Modify one or more sequences to include Insertions or Deletions
}
\details{
Takes an \link[Biostrings:XString-class]{Biostrings::XString} or \link[Biostrings:XStringSet-class]{Biostrings::XStringSet} object and moddfies
the sequence to incorporate InDels.
The expected types of data determine the behaviour, with the following
expectations describing how the function will incorporate data\tabular{llll}{
   Input Data Type \tab Exons Required \tab Use Case \tab Returned \cr
   XString \tab Y \tab Modify a Reference Transcriptome \tab XString \cr
   XStringSet \tab Y \tab Modify a Reference Transcriptome \tab XStringSet \cr
   DNAStringSet \tab N \tab Modify a Reference Genome \tab DNAStringSet \cr
   BSgenome \tab N \tab Modify a Reference Genome \tab DNAStringSet \cr
}
}
