% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subSNP.R
\name{subSNP}
\alias{subSNP}
\title{Insert SNPs into a XStringSet}
\usage{
subSNP(seq, snps, alt_col = "ALT", ...)
}
\arguments{
\item{seq}{A BSgenome, DNAStringSet, RNAStringSet or other XStringSet.}

\item{snps}{A GRanges object with SNP positions and a column containing
the alternate allele}

\item{alt_col}{Column name in the mcols element of \code{snps} containing the
alternate allele}

\item{...}{Passed to \code{\link[Biostrings:replaceLetterAt]{Biostrings::replaceLetterAt()}}}
}
\value{
An object of the same class as the original object, but with SNPs
inserted at the supplied positions
}
\description{
Modify a set of sequences using SNPs
}
\details{
If providing a BSgenome object, this will first be coerced to a DNAStringSet
which can be time consuming#'
}
\examples{
library(GenomicRanges)
seq <- DNAStringSet(c(chr1 = "AAGC"))
snps <- GRanges("chr1:2")
snps$ALT <- "G"
snps
seq
subSNP(seq, snps)

}
