% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomogrify.R
\name{genomogrify}
\alias{genomogrify}
\alias{genomogrify,XStringSet,GRanges-method}
\alias{genomogrify-methods}
\alias{genomogrify,BSgenome,GRanges-method}
\alias{genomogrify,BSgenome,VcfFile-method}
\alias{genomogrify,XStringSet,VcfFile-method}
\title{Mogrify a genome using a set of variants}
\usage{
genomogrify(x, var, ...)

\S4method{genomogrify}{XStringSet,GRanges}(
  x,
  var,
  alt_col = "ALT",
  tag = NULL,
  sep = "_",
  var_tags = FALSE,
  var_sep = "_",
  ...
)

\S4method{genomogrify}{BSgenome,GRanges}(
  x,
  var,
  alt_col = "ALT",
  names,
  tag = NULL,
  sep = "_",
  var_tags = FALSE,
  var_sep = "_",
  verbose = TRUE,
  ...
)

\S4method{genomogrify}{BSgenome,VcfFile}(
  x,
  var,
  alt_col = "ALT",
  names,
  tag = NULL,
  sep = "_",
  var_tags = FALSE,
  var_sep = "_",
  which,
  verbose = TRUE,
  ...
)

\S4method{genomogrify}{XStringSet,VcfFile}(
  x,
  var,
  alt_col = "ALT",
  tag = NULL,
  sep = "_",
  var_tags = FALSE,
  var_sep = "_",
  which,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{A DNAStringSet or BSgenome}

\item{var}{GRanges object containing the variants, or a
\link[VariantAnnotation:VcfFile-class]{VariantAnnotation::VcfFile}}

\item{...}{Passed to \link[parallel:mclapply]{parallel::mclapply}}

\item{alt_col}{The name of the column with \code{var} containing alternate bases}

\item{tag}{Optional tag to add to all sequence names which were modified}

\item{sep}{Separator to place between seqnames names & tag}

\item{var_tags}{logical(1) Add tags indicating which type of variant were
incorporated, with 's', 'i' and 'd' representing SNPs, Insertions and
Deletions respectively}

\item{var_sep}{Separator between any previous tags and variant tags}

\item{names}{Sequence names to be mogrified}

\item{verbose}{logical(1) Print progress messages while running}

\item{which}{GRanges object passed to \link[VariantAnnotation:ScanVcfParam-class]{VariantAnnotation::ScanVcfParam} if
using a VCF directly}
}
\value{
XStringSet with variant modified sequences
}
\description{
Use a set of SNPS, insertions and deletions to modify a reference genome
}
\examples{
library(GenomicRanges)
dna <- DNAStringSet(c(chr1 = "ACGT", chr2 = "AATTT"))
var <- GRanges(c("chr1:1", "chr1:3", "chr2:1-3"))
var$ALT <- c("C", "GG", "A")
dna
genomogrify(dna, var)
genomogrify(dna, var, tag = "mod")
genomogrify(dna, var, var_tags = TRUE)


}
